+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2025.2.3]
 
+### `xDebugger` enhancements
+- Create type renderer: The item type is not present in the *items.xml files [#1531](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1531)
+
 ### `Other` enhancements
 - Migrated to immutable Settings [#1527](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1527)